TimeSeries/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
wandb>=0.15.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
joblib>=1.0.0

# Time Series Specific Libraries
darts>=0.27.0
sktime>=0.24.0
statsmodels>=0.14.0
torch>=2.0.0
prophet>=1.1.0
xgboost>=2.0.0
lightgbm>=4.0.0